Spit Integrate Test
